+++ CHANGELOG.md
@@ -69,9 +69,12 @@
 - FEAT: starting work on modules
 - FEAT: call public functions from other modules
 - FEAT: access public classes from other modules
+- FEAT: access global vars from other modules
 
 ## TODO
 
+- access enums from other modules
+- Declare types out of scope
 - Add in for loops 
 - Moduling system
 - References?